api ok

+++ app/src/main/java/com/train/train_manager/act/ruku_add/RukuAddAdapter.java
@@ -67,7 +67,7 @@ public class RukuAddAdapter extends BaseAdapter {
         holder.info1.setText("BST物料编码：" + bean.bstPartNo);
         holder.info2.setText("入库库位：" + bean.tLocation);
         holder.info3.setText("入库时间：" + bean.operTime);
-        holder.info.setText(bean.qty + bean.ume);
+        holder.info.setText((int)bean.qty + bean.ume);
         return convertView;
     }
 }
